Refactored Pagination

+++ elide-core/src/main/java/com/yahoo/elide/core/datastore/inmemory/InMemoryStoreTransaction.java
@@ -330,7 +330,7 @@ public class InMemoryStoreTransaction implements DataStoreTransaction {
         }
 
         if (pagination.returnPageTotals()) {
-            pagination.setPageTotals(records.size());
+            pagination.setPageTotals((long) records.size());
         }
         return records.subList(offset, endIdx);
     }
